Introduce feature `break`
The `Break` expression constructor of the AST was gated only by the
feature `loop`. This commit introduces the feature `break` and makes
the `Break` constructor gated by the witness tuple `break * loop`.

Thanks to that, the phase `Cf_into_monads` now can remove
`Break`-nodes in a type-safe manner. (though, for now, it removes
`Break`s in favor of a runtime error, see issue #96)