Remove `ExtendedDefPathItem`, instead have a map from `DefId`s to impl details
The design of `ExtendedDefPathItem` was a bit clunky. The original
idea was to simply have one `DefPathItem` type augmented with
informations on the `Impl` constructor (what type is implemented +
with what trait constraints). Sadly @sonmarcho found out this design
could lead to mutually refrencing `DefId`s.

`ExtendedDefPathItem` results from this observation:
``ExtendedDefPathItem`'s `Impl` variant refers to non-extended
`DefPathItem`s, thus breaking the cycle.

But this is quite heavy, and requires us, in the frontend, to *decide*
where a `DefÌd` should be extended or not. In other words, we needed
to decide when we give full detail about `Impl`s in `DefId`s. This was
conflicting with the "give as much info as possible" motto of the
exporter.

Instead, this commit makes two changes:

 - add a `DefId` mutable vector in the state, that collects every
   single `DefId` we ever translate --note this is also useful for
   detecting external defintions in Charon/Aeneas (and probably in Hax
   as well at some point);

 - define a `impl_def_ids_to_impled_types_and_bounds` function on
   states that provides a mapping from `Impl` `DefId`s to their
   implemented types and bounds.

Instead of inlining directly this information on each `DefId`, the
consumer of the output of Hax's frontend will need to lookup the
information in a separate map.