feat(front): into. a special core extraction mode that skips attrs & bodies
Rust's core library is quite big, and is nicely documented. Currently,
Hax frontend exports each item with its attributes and the attributes
of each of its parent. Thus, there is a lot of duplication.

The JSON outputed for core is 11gb.

This commit introduces the env var `HAX_CORE_EXTRACTION_MODE`. When
set to `on`, the frontend skip attributes and writes `()` in the
bodies.

This makes the JSON drop from 11gb to 500mb.

TODO: make the bodies `panic!()` instead of `()`.