Add `--hax-engine-path`, auto JS download, introduce `RUST_SRC_HASH`
Introduces `RUST_SRC_HASH`, the hash of the frontend's rust
sources. This is useful to assert datatype compatibility between the
engine and the frontend. The frontend now checks for compatibility.

`RUST_SRC_HASH` is also helpful in downloading a compatible
engine. When the user consents to downloading Hax engine from the
internet, we simply look for any engine with a compatible
`RUST_SRC_HASH`.

Note: when we will have proper releases and versioning, release
versions will be preferred over `RUST_SRC_HASH` for downloading Hax
Engine.

The commit also introduces a const `DISTS_ROOT_URI` in `cargo_hax.rs`:
this refers to `[https://hacspec.org/hax-engine-dists/dists`](https://hacspec.org/hax-engine-dists/dists%60), that
doesn't exist yet. This is a placeholder for a directory where should
be uploaded JS prebuilt hax engines with specific filenames reflecting
the `RUST_SRC_HASH` they were built with. An easy way of doing this is
just creating a `hax-engine-dists` repo under the `hacspec` org, and
then have GH actions to populate it at each commit on
`hacspec-v2/main`.