feat: opts for F* backend (`z3rlimit`, `fuel`, `ifuel`)
This commit adds options for the F* backend to control F*'s
`z3rlimit`, `fuel` and `ifuel`.

It also add an `open FStar.Mul` in the header of extracted modules,
which is often convenient when playing with extracted modules.

Last but not least: this commits adds a newline at the end of the file
🎉🥳✨ (this always hurts my eyes in diffs)