fix(engine/import-thir): fix leaking errors
The `capture` function of `diagnostics.ml` was baldy implemented.
Seems like in the expression `(f (), g ())` OCaml first runs `g` and
then `f`!

I'm using a state to capture errors, and using `(f (), !state)`, I was
ignoring the side effects of `f`, which is... the only point of that
`capture` function.