fix(engine): drop visitors, x4 compile time speedup 🎉🎉🎉🎉🎉
I finally found out why the compilation of the engine of Hax was so
damn slow!

Everything is functorized in the engine, and we instantiate new ASTs
(with the functor `AST.Make`) all the time.

The catch is that the `visitors` library generates an a LOT of code,
and cannot generate interfaces. We had no way of decoupling visitor
and the types defining the AST. Each time we instantiate an AST with a
feature set, we also instantiate the huge visitors.

The performance issue was comming from there, both in term of time and
memory!

This commit removes visitors entirely, and replace them with a
`ast_visitors.ml` file that I generated by hacking a small
script. Ideally, I would like this file to be generated, but currently
I just check in a base, and we will tweak it when we change the AST.

The engine now compiles in 40 seconds on my laptop! (instead of 2:40 minutes)

This commit fixes (for good) the issue #197!