fix(engine): allow default generics
Default generics are just a nicety provided by Rust, but at the end,
Rustc always inline the default when not provided. We don't want to
propagate this nicety in the backends, since it's useful only to ease
writing new code.

Thus, this PR just drops the default information if any.

Fixes #596