feat(exporter): drop clause ids, use `PredicateId` both on `Predicate`s and `Clause`s
In commit [fca56a], the type `Clause` has been renamed as
`ClauseKind`, then commit [21226e] makes `Clause` a
`Binder<ClauseKind>`, essentially (with an indirection via
`PredicateKind` and an invariant where the predicate kind has to be a
`ty::PredicateKind::Clause(ClauseKind)`).

That means that now, Rustc always provides `Clause` with
binders. Thus, it now makes sense to systematically compute clause
identifiers with binders.

Since clauses are a subset of predicates, we now always lift clauses
to predicates before computing identifiers. This was not possible
before, since we lacked binder information to properly lift clauses to
predicates

[fca56a]: https://github.com/rust-lang/rust/commit/fca56a8d2c6f803aeec51167dd4dd22529f8bdd7
[21226e]: https://github.com/rust-lang/rust/commit/21226eefb2e2465b9730c37e7d08655865610d90