fix(exporter): improve `predicate_equality`
When `try_normalize_erasing_regions` fails, regions are not
erased. When comparing predicates, regions never matter: we need to
make sure they are erased (an erased lifetime will show up as `&`
instead of `&'named`).

Also, erased regions still show up in the binder: whence we skip it.

This commit does a last thing: instead of dealing with constness by
replacing in strings, we call `without_const`.