fix(proof-libs): fix implicits
Since we are now printing types explicitely, the F* library has no more freedom in term of implicits.

For example, operations on vectors always take two implicits: the type
of the items held by the vector, but also the type of the
allocator. In F*, we don't care about the allocator, thus, we model
any allocator with unit. In this setting, taking an implicit allocator
each time we take a vector is useless in F*. But now, the extraction gives (as an implicit) the allocator, so we need to expect an allocator.