flake: Add rust-overlay as a separate input
My reasoning is as follows: we need an up-to-date rust-overlay in order
to access recent nightlies. Currently we follow crane's rust-overlay,
but recent versions of crane don't have a rust-overlay input anymore. So
we'll need to have rust-overlay as input like this eventually anyway.
Hence I did this without even updating crane because there was no need.