fix(frontend): kill `crate_type` in `HaxMeta`
This commit removes the field `crate_type` in the struct `HaxMeta`,
and changes the path of `haxmeta` files.

Originally, my intention was to name the `haxmeta` files very closely
to `rmeta` files, but the `haxmeta` don't have anything to do with
crate types.

Also, the code was assuming a crate would have only one type, which is
false, as the type suggests... That was making the frontend panic on
certain crates.