fix(backend/fstar): pre/post are `Type0` predicates, not boolean predicates
For a method `f` in a trait, we generate two fields: `f_pre` and
`f_post`. Those used to be boolean predicates, because we inherited
the type from Rust. This commit changes that: we now enforce `f_pre`
and `f_post` to be arrow type with `Type0` as last component.