docs(book): depend on `hax-lib` only when using hax
Adding `hax-lib` as a non-`cfg`-gated dependency resulted in extra
dependencies being added to the dependency tree.
Some of these transitive dependencies also required the `alloc` crate,
and this setup could thus not be used as-is in no-alloc projects.

This change makes it so that the dependency is only present when the
`hax` `cfg` flag is present, solving these issues.