fix(engine/fstar): fix super typeclasses attributes
The PR #902 was reverted by #909 because #902 was generated redundant
fields names in classes for super clauses, which is a problem for F*.

This PR changes the naming scheme of super clauses for typeclasses in
F*. Those names now take into account the trait name. The super trait
names are now a 5 characters base 62 encoding of the hash of the
predicate ID concatenated with the trait name.

Fixes #630.