fix(engine): make consts fun. calls of arity 0 when there is an impl expr
The engine was dropping trait information when importing constants:
now, if a constant `T::CONST` is an associated item of a non-inherent
trait, we represent it `T::CONST` as a call to `CONST()` with the
implementation expression for `T`.

Note that, in the engine, constants are already represented as
functions of arity zero. Rust functions of arity zero are represented
as function that take exactly one `unit` input.

Fixes #840.