feat(engine/f*): add dummy support for floats
This PR makes the F* backend accept floats.
This will not typecheck at the end, since there is no model of floats in F*.

The next step is to add a very opaque model of floats to F*.

This is useful for putting rust by examples on CI: a lot of examples are using floats.