fix(engine/deps): preserve definitions order whenever possible
In most of the backends, a top-level definition should come prior to it's use.
This is not true in Rust.

Thus, we need to sort the items so that item `a` comes before it is
used in another item `b`.

We used to do that with a topological sort with the ocamlgraph
library. But since this topological sort operated on a graph, we were
loosing the original order of items.

This commit now:
 - construct a the transtive closure of the dependency graph
 - use this closure to stable sort the items

The comparison function is defined as follows (`g` denotes the transitive closure):
 1. if there is an edge `a ⟶ b` in `g` and an edge `b ⟶ a`, we are in a recursive bundle, hence we return `0`: `a` equals `b`
 2. if there is an edge `a ⟶ b` in `g`, we return `-1`
 2. if there is an edge `b ⟶ a` in `g`, we return `1`
 2. otherwise there is no relation between `a` and `b`, we return `0`